Refactor MC test code into single file
Refs #15887
