+++ Framework/Algorithms/test/MCInteractionVolumeTest.h
@@ -4,14 +4,7 @@
 #include <cxxtest/TestSuite.h>
 
 #include "MantidAlgorithms/SampleCorrections/MCInteractionVolume.h"
-#include "MantidAPI/Sample.h"
-#include "MantidAPI/SampleEnvironment.h"
-#include "MantidGeometry/Objects/Object.h"
-#include "MantidGeometry/Objects/ShapeFactory.h"
-#include "MantidKernel/Material.h"
-#include "MantidKernel/PseudoRandomNumberGenerator.h"
-
-#include "MantidTestHelpers/ComponentCreationHelper.h"
+#include "MonteCarloTesting.h"
 
 #include <gmock/gmock.h>
 
@@ -31,6 +24,7 @@ public:
   //----------------------------------------------------------------------------
   void test_Absorption_In_Solid_Sample_Gives_Expected_Answer() {
     using Mantid::Kernel::V3D;
+    using namespace MonteCarloTesting;
     using namespace ::testing;
 
     // Testing inputs
@@ -50,6 +44,7 @@ public:
 
   void test_Absorption_In_Sample_With_Hole_Can_Scatter_In_All_Segments() {
     using Mantid::Kernel::V3D;
+    using namespace MonteCarloTesting;
     using namespace ::testing;
 
     // Testing inputs
@@ -80,6 +75,7 @@ public:
 
   void test_Absorption_In_Sample_And_Environment_Can_Scatter_In_All_Segments() {
     using Mantid::Kernel::V3D;
+    using namespace MonteCarloTesting;
     using namespace ::testing;
 
     // Testing inputs
@@ -135,6 +131,7 @@ public:
 
   void test_Track_With_Zero_Intersections_Throws_Error() {
     using Mantid::Kernel::V3D;
+    using namespace MonteCarloTesting;
     using namespace ::testing;
 
     // Testing inputs
@@ -152,97 +149,6 @@ public:
                      std::logic_error);
   }
 
-  //----------------------------------------------------------------------------
-  // Non-test methods
-  //----------------------------------------------------------------------------
-private:
-  enum class TestSampleType { SolidSphere, Annulus, SamplePlusCan };
-
-  class MockRNG final : public Mantid::Kernel::PseudoRandomNumberGenerator {
-  public:
-    MOCK_METHOD0(nextValue, double());
-    MOCK_METHOD2(nextValue, double(double, double));
-    MOCK_METHOD2(nextInt, int(int, int));
-    MOCK_METHOD0(restart, void());
-    MOCK_METHOD0(save, void());
-    MOCK_METHOD0(restore, void());
-    MOCK_METHOD1(setSeed, void(size_t));
-    MOCK_METHOD2(setRange, void(const double, const double));
-  };
-
-  Mantid::API::Sample createTestSample(TestSampleType sampleType) {
-    using Mantid::API::Sample;
-    using Mantid::Kernel::Material;
-    using Mantid::Kernel::V3D;
-    using Mantid::Geometry::Object_sptr;
-    using Mantid::PhysicalConstants::getNeutronAtom;
-
-    using namespace Mantid::Geometry;
-
-    Sample testSample;
-    if (sampleType == TestSampleType::SamplePlusCan) {
-      testSample = createSamplePlusCan();
-    } else {
-      Object_sptr shape;
-      if (sampleType == TestSampleType::SolidSphere) {
-        shape = ComponentCreationHelper::createSphere(0.1);
-      } else if (sampleType == TestSampleType::Annulus) {
-        shape = createAnnulus(0.1, 0.15, 0.15, V3D(0, 0, 1));
-      } else {
-        throw std::invalid_argument("Unknown testing shape type requested");
-      }
-      shape->setMaterial(Material("Vanadium", getNeutronAtom(23), 0.02));
-      testSample.setShape(*shape);
-    }
-    return testSample;
-  }
-
-  Mantid::API::Sample createSamplePlusCan() {
-    using Mantid::API::Sample;
-    using Mantid::API::SampleEnvironment;
-    using Mantid::Kernel::Material;
-    using Mantid::Kernel::V3D;
-    using Mantid::PhysicalConstants::getNeutronAtom;
-
-    // Create an annulus Vanadium can with silicon sample
-    const double height(0.05), innerRadius(0.0046), outerRadius(0.005);
-    const V3D centre(0, 0, -0.5 * height), upAxis(0, 0, 1);
-    // Can
-    auto environment =
-        Mantid::Kernel::make_unique<SampleEnvironment>("Annulus Can");
-    auto can = createAnnulus(innerRadius, outerRadius, height, upAxis);
-    can->setMaterial(Material("Vanadium", getNeutronAtom(23), 0.02));
-    environment->add(*can);
-    // Sample volume
-    auto sampleCell = ComponentCreationHelper::createCappedCylinder(
-        innerRadius, height, centre, upAxis, "sample");
-    sampleCell->setMaterial(Material("Si", getNeutronAtom(14), 0.15));
-
-    // Sample object
-    Sample testSample;
-    testSample.setShape(*sampleCell);
-    testSample.setEnvironment(environment.release());
-    return testSample;
-  }
-
-  Mantid::Geometry::Object_sptr
-  createAnnulus(double innerRadius, double outerRadius, double height,
-                const Mantid::Kernel::V3D &upAxis) const {
-    using Mantid::Geometry::ShapeFactory;
-    using Mantid::Kernel::V3D;
-
-    // Cylinders oriented along up, with origin at centre of cylinder
-    const V3D centre(0, 0, -0.5 * height);
-    const std::string inner = ComponentCreationHelper::cappedCylinderXML(
-        innerRadius, height, centre, upAxis, "inner");
-    const std::string outer = ComponentCreationHelper::cappedCylinderXML(
-        outerRadius, height, centre, upAxis, "outer");
-
-    // Combine shapes
-    std::ostringstream os;
-    os << inner << outer << "<algebra val=\"(outer (# inner))\" />";
-    return ShapeFactory().createShape(os.str());
-  }
 };
 
 #endif /* MANTID_ALGORITHMS_MCINTERACTIONVOLUMETEST_H_ */
